Customizations on Mailbox Plans don't persist?
This morning I was informed that that any build changes or deployments could reset mailbox plans.  Customization (e.g. - lower mailbox storage quotas) is not guaranteed to be retained or carried forward.  I have a customer case open right now that I'm trying to resolve... If we confirm this is by design, this is counter intuitive and confusing for customers, engineers/consultants... we need to update this article accordingly or prod the PG to change this...